requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.2
django-crispy-forms==1.8.1
django-paypal==1.0.0
idna==2.8
lxml==4.4.2
Pillow==5.4.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tlslite-ng==0.7.5
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
